Fixed warnings and a few minor bugs when compiling on a different architecture.
git-svn-id: http://yate.null.ro/svn/yate/trunk@2724 acf43c95-373e-0410-b603-e72c3f656dc1

+++ engine/Socket.cpp
@@ -609,7 +609,7 @@ bool File::setBlocking(bool block)
 #else
     unsigned long flags = 1;
     flags = ::fcntl(m_handle,F_GETFL);
-    if (flags < 0) {
+    if ((signed long)flags < 0) {
 	copyError();
 	return false;
     }
@@ -1577,7 +1577,7 @@ bool Socket::setBlocking(bool block)
     return checkError(::ioctlsocket(m_handle,FIONBIO,(unsigned long *) &flags));
 #else
     flags = ::fcntl(m_handle,F_GETFL);
-    if (flags < 0) {
+    if ((signed long)flags < 0) {
 	copyError();
 	return false;
     }
